requirements.txt
gitpython
dateutils
beautifulsoup4
requests